Refactoring main_screen pages
