+++ lib/routes.dart
@@ -18,6 +18,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:picos/screens/legals_screen.dart';
 import 'package:picos/screens/login_screen.dart';
+import 'package:picos/screens/main_screen/overview_screen/overview.dart';
 import 'package:picos/screens/my_medications_screen/add_medication_screen.dart';
 import 'package:picos/screens/my_therapies_screen/add_therapy_screen.dart';
 import 'package:picos/screens/my_therapies_screen/my_therapies_screen.dart';
@@ -32,7 +33,6 @@ import 'package:picos/screens/study_nurse_screen/configuration_screen/configurat
 import 'package:picos/screens/study_nurse_screen/configuration_screen/configuration_screen.dart';
 import 'package:picos/screens/main_screen/bottom_bar.dart';
 import 'package:picos/screens/my_medications_screen/my_medications_screen.dart';
-import 'package:picos/screens/overview_screen/overview_screen.dart';
 import 'package:picos/screens/physician_list_screen/physicians_screen.dart';
 import 'package:picos/screens/study_nurse_screen/study_nurse_screen.dart';
 import 'package:picos/screens/my_therapies_screen/view_therapy_screen.dart';
@@ -64,7 +64,7 @@ class Routes {
           const ConfigurationPages(),
       '/study-nurse-screen/configuration-finish-screen': (BuildContext ctx) =>
           const ConfigurationFinishedScreen(),
-      '/overview-screen/overview': (BuildContext ctx) => const OverviewScreen(),
+      '/overview-screen/overview': (BuildContext ctx) => const Overview(),
       '/my-therapy-screen/add-therapy': (BuildContext ctx) =>
           const AddTherapyScreen(),
       '/my-therapy-screen/view-therapy': (BuildContext ctx) =>
