+++ lib/screens/main_screen/overview_screen/widgets/contact_section.dart
@@ -18,7 +18,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:picos/screens/overview_screen/widgets/section.dart';
+import 'package:picos/screens/main_screen/overview_screen/widgets/section.dart';
 
 /// Widget which shows a graph
 class ContactSection extends StatelessWidget {
