+++ lib/screens/main_screen/overview_screen/widgets/my_health_section.dart
@@ -18,8 +18,8 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:picos/screens/overview_screen/widgets/tile.dart';
-import 'package:picos/screens/overview_screen/widgets/section.dart';
+import 'package:picos/screens/main_screen/overview_screen/widgets/section.dart';
+import 'package:picos/screens/main_screen/overview_screen/widgets/tile.dart';
 
 /// Widget which displays health-related information
 class MyHealthSection extends StatelessWidget {
